tests/requirements.txt
# Packages needed to run the tests.
# Switch into a virtual environment
# pip install -r requirements.txt

dash[dev,testing]>=1.15.0
Pillow
-e .
-r ./demos/usage-algorithm/requirements.txt